Adjust implementation to use float instead of int for heat values

+++ satviz-consumer-native/include/satviz/Bindings.h
@@ -17,7 +17,7 @@ typedef struct CWeightUpdate {
 typedef struct CHeatUpdate {
   unsigned int n;
   int *index;
-  int *heat;
+  float *heat;
 } CHeatUpdate;
 
 // Graph
